+++ .github/workflows/copilot-setup-steps.yml
@@ -46,7 +46,7 @@ jobs:
             ~/.cargo/registry/cache/
             ~/.cargo/git/db/
             packages/biliass/rust/target/
-            packages/yutto-core/rust/target/
+            rust/target/
           key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
 
       - name: Install tools
