+++ .github/workflows/release.yml
@@ -6,84 +6,35 @@ on:
       - "v*" # Push events to matching v*, i.e. v1.0, v20.15.10
   workflow_dispatch:
 
+permissions:
+  contents: read
+
 jobs:
-  release-build:
+  msrv:
+    name: Rust 1.85 compatibility
     runs-on: ubuntu-latest
-
     steps:
-      - name: Checkout
-        uses: actions/checkout@v7
-
-      - name: Install uv
-        uses: astral-sh/setup-uv@v9.0.0
+      - uses: actions/checkout@v7
+      - name: Install Rust 1.85 toolchain
+        uses: dtolnay/rust-toolchain@1.85.0
         with:
-          enable-cache: true
-
-      - name: Install python
-        uses: actions/setup-python@v7
-        with:
-          python-version: "3.x"
-
-      - name: Install Rust toolchain
-        uses: dtolnay/rust-toolchain@stable
-
-      - name: Install just
-        uses: extractions/setup-just@v4
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: build release distributions
+          components: clippy
+      - name: Check MSRV compatibility
+        working-directory: rust
         run: |
-          just build
+          cargo check --locked --all-targets --all-features
+          cargo clippy --locked --all-targets --all-features -- -D warnings
+          cargo test --locked --all-targets --all-features
 
-      - name: upload dists
-        uses: actions/upload-artifact@v7
-        with:
-          name: release-dists
-          path: dist/
-
-  publish-pypi:
-    runs-on: ubuntu-latest
-    name: Publish to PyPI
-    if: "startsWith(github.ref, 'refs/tags/')"
-    needs:
-      - release-build
-    permissions:
-      id-token: write
-
-    steps:
-      - name: Retrieve release distributions
-        uses: actions/download-artifact@v8
-        with:
-          name: release-dists
-          path: dist/
-
-      - name: Install uv
-        uses: astral-sh/setup-uv@v9.0.0
-        with:
-          enable-cache: true
-
-      - name: Publish release distributions to PyPI
-        run: uv publish -v
-
-  publish-release:
-    runs-on: ubuntu-latest
-    name: Publish to GitHub
-    if: "startsWith(github.ref, 'refs/tags/')"
-    needs:
-      - release-build
+  build-and-release:
+    needs: msrv
+    uses: ./.github/workflows/native-build-and-release.yml
+    with:
+      working-directory: .
+      artifact-prefix: yutto
+      release-ref-prefix: refs/tags/v
+      # ring's aarch64 assembly does not compile with the manylinux2014 cross toolchain.
+      manylinux-aarch64: "2_28"
     permissions:
       contents: write
-    steps:
-      - uses: actions/download-artifact@v8
-        with:
-          name: release-dists
-          path: dist/
-      - name: Get tag name
-        run: echo "RELEASE_VERSION=${GITHUB_REF#refs/*/}" >> $GITHUB_ENV
-      - name: Publish to GitHub
-        uses: softprops/action-gh-release@v3
-        with:
-          draft: true
-          files: dist/*
-          tag_name: ${{ env.RELEASE_VERSION }}
+      id-token: write
