--- .github/workflows/yutto-core-build-and-release.yml
@@ -1,57 +1,0 @@
-name: Build and Release (yutto-core)
-
-on:
-  push:
-    tags:
-      - "yutto-core@*"
-    branches:
-      - main
-    paths:
-      - "packages/yutto-core/**"
-      - ".github/workflows/yutto-core-build-and-release.yml"
-      - ".github/workflows/native-build-and-release.yml"
-  pull_request:
-    paths:
-      - "packages/yutto-core/**"
-      - "src/yutto/downloader/**"
-      - ".github/workflows/yutto-core-build-and-release.yml"
-      - ".github/workflows/native-build-and-release.yml"
-  merge_group:
-  workflow_dispatch:
-
-permissions:
-  contents: read
-
-concurrency:
-  group: ${{ github.workflow }}-${{ github.ref }}
-  cancel-in-progress: true
-
-jobs:
-  msrv:
-    name: Rust 1.85 compatibility
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v7
-      - name: Install Rust 1.85 toolchain
-        uses: dtolnay/rust-toolchain@1.85.0
-        with:
-          components: clippy
-      - name: Check MSRV compatibility
-        working-directory: packages/yutto-core/rust
-        run: |
-          cargo check --locked --all-targets --all-features
-          cargo clippy --locked --all-targets --all-features -- -D warnings
-          cargo test --locked --all-targets --all-features
-
-  build-and-release:
-    needs: msrv
-    uses: ./.github/workflows/native-build-and-release.yml
-    with:
-      working-directory: packages/yutto-core
-      artifact-prefix: yutto-core
-      release-ref-prefix: refs/tags/yutto-core@
-      # ring's aarch64 assembly does not compile with the manylinux2014 cross toolchain.
-      manylinux-aarch64: "2_28"
-    permissions:
-      contents: write
-      id-token: write
