+++ justfile
@@ -2,7 +2,6 @@ set positional-arguments
 
 VERSION := `uv run scripts/get-version.py src/yutto/__version__.py`
 BILIASS_VERSION := `uv run scripts/get-version.py packages/biliass/src/biliass/__version__.py`
-YUTTO_CORE_VERSION := `uv version --package yutto-core --short`
 DOCKER_NAME := "siguremo/yutto"
 
 run *ARGS:
@@ -19,21 +18,12 @@ fmt:
   uv run ruff format .
 
 lint:
-  uv run ty check --error-on-warning src/yutto packages/biliass/src/biliass packages/yutto-core/src/yutto_core tests
+  uv run ty check --error-on-warning src/yutto packages/biliass/src/biliass tests
   uv run ruff check .
   uv run typos
 
 build:
   uv build
-
-build-yutto-core:
-  uv build --package yutto-core
-
-release-yutto-core:
-  @echo 'Tagging yutto-core@{{YUTTO_CORE_VERSION}}...'
-  git tag "yutto-core@{{YUTTO_CORE_VERSION}}"
-  @echo 'Push to GitHub to trigger publish process...'
-  git push --tags
 
 release:
   @echo 'Tagging v{{VERSION}}...'
