--- packages/yutto-core/pyproject.toml
@@ -1,16 +1,0 @@
-[project]
-name = "yutto-core"
-version = "0.1.0"
-description = "Rust core for yutto"
-requires-python = ">=3.11"
-license = "GPL-3.0-only"
-dependencies = []
-
-[tool.maturin]
-features = ["pyo3/extension-module"]
-manifest-path = "rust/crates/yutto/Cargo.toml"
-module-name = "yutto_core._core"
-
-[build-system]
-requires = ["maturin>=1.4,<2.0"]
-build-backend = "maturin"
