--- packages/yutto-core/src/yutto_core/py.typed
@@ -1,1 +1,0 @@
-
