+++ rust/crates/haya-http/examples/download.rs
