+++ rust/crates/haya-http/tests/http_download.rs
