+++ rust/crates/haya/src/buffer.rs
