+++ rust/crates/haya/src/sink.rs
