+++ src/yutto/_core.pyi
